modules/news_bot/pyproject.toml
[tool.poetry]
name = "news-bot"
version = "0.1.0"
description = "French News Bot (Falcon7B-LLM) avec LangChain"
authors = ["Angnami <mhtgoudjemi@gmail.com>"]
readme = "README.md"
packages = [{include = "news_bot"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
qdrant-client = "^1.7.2"
langchain = "^0.1.4"
comet-ml = "^3.37.0"
transformers = "^4.37.2"
peft = "^0.8.1"
accelerate = "^0.26.1"
python-dotenv = "^1.0.1"
scipy = "^1.12.0"
einops = "^0.7.0"
comet-llm = "^2.1.1"
fire = "^0.5.0"
bitsandbytes = "^0.42.0"
torch = "^2.2.0"
streamlit = "^1.30.0"
unstructured = "0.12.0"
beam-sdk = "^0.15.13"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
ruff = "^0.1.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.isort]
case-sensitive = true